osemgrep: Cleanup Semgrep_settings.yml and pass ~token_opt more
Not a huge fan of this lazy trick to get the semgrep settings.
In any case I think it's better to be explicit that rely on
this global and so passing around a token seems better,
in which case there's no point anymore in optimizing Semgrep_settings
using Lazy.

test plan:
make osempass