cleanup: Remove Bloom filter optimization
Follows: 2428d17ee7c ("Disable Bloom filter optimization by default (#6197)")

It was probably more impactful when it was first introduced, but the
engine has evolved and nowadays it does not seem to have a noticeable
impact on performance. We disabled it by default on Sep 2022 and this
went completely unnoticed.

What triggered this was that the Bloom filter had tricky interactions
with const-prop and sym-prop, see #4670, and PA-1920 (or PR #6179).

test plan:
make test