more js updates (#7775)
- Build CommonJS and ES module flavors of the engine (thought I could
get away with just CJS, but we need ESM for dynamic imports in the
frontend)
- I was mistaken when I thought that we could simply load parsers for
whatever languages are listed in the rules JSON. Some languages depend
on multiple parsers (e.g. `dockerfile` needs `bash` -- my initial
workaround was to simply include bash in the dockerfile parser), and now
I'm realizing that `metadata-comparison` needs `python`. Instead of
playing whack-a-mole, let's surface missing languages in the engine so
that we can load then and then retry
- Move engine tests into tests folder

test plan:
- `make test`