osemgrep: NEW incremental display of match results in text mode
Note that it works only for -j 1 for now, which is seriously limited.
The issue is that the file_match_result_hook is run from child process
ran via Parmap and so we need to find a way to synchronize those calls
to serialize the use of stdout and not get interwinded output.

Any suggestions?

test plan:
$ osemgrep --legacy -j 1 --no-ast-caching -l ocaml -e ': Common.filename' .
display the matches all at once

$ osemgrep -j 1 --no-ast-caching -l ocaml -e ': Common.filename' .
display its matches gradually.