fix: Introduce type inference using Type.ml
Here, I bring `Type.ml` over from the Pro Engine, and use it to it to
implement a couple cases of type inference, just enough to showcase the
idea. I don't think that `AST_generic.type_` is suitable for use as a
type representation for any sophisticated type inference.

This way, we can build out local type inference in OSS, improving
matching for typed metavariables and possibly other things when local
type inference is able to determine a type. We can also make use of the
Pro Engine's resolved types for symbols defined outside of the local
scope to further improve typed metavariable matching in the Pro Engine.

Test plan: Automated tests