fix: Handle variables in type inference
Out of the two new test cases, previously only the first one matched. We
already handled the special case where a typed metavariable was matching
a variable with a resolved type. However, the second case did not match.

In many cases we can assume that the type of a variable has already been
resolved. This is handled by `Naming_AST.ml` (and the equivalent in Pro
Engine). In some languages, type annotations do not exist on all
variables, so we might need to recurse on the `id_svalue` if it has one.
For now, we assume that naming has been able to resolve the type of a
variable.

Test plan: Automated tests