fix(perf): set better limits on analyze_rule
Closes https://github.com/returntocorp/semgrep/issues/7839

In Analyze_rule, we distribute out formulas like `(x0 ^ x1) v (y0 ^ y1)` into cnf.
Let's call `x0 ^ x1` p and `y0 ^ y1` q. Currently, we check that `p` and `q` are
individually under 1,000,000 elements to prevent memory blowup. However, `p` and `q`
produce a formula that is `p * q` elements long. So, even under these constraints,
if `p` and `q` are both large, this step can still take too long.

Instead, what we need to do is guard the result. This PR currently requires that
that result be under 10,000,000 elements. It leaves the `p` and `q` restrictions
intact just in case. This allows us to take advantage of `List.compare_length`
in case we get an incredibly long `p` or `q`.

To consider: do we still need the gates for `p` and `q`? Is it possible to get
a list for `p` or `q` so large that traversing it is a problem?

Test plan: will add tests