sym-prop: Allow some Sym-svalues to survive after JOINs
Previously any Sym-value would be "killed" at a JOIN point, as a
precautionary measure to avoid `id_svalue` cycles. Here we relax
this condition to allow trivial cases to work. Basically, if a
Sym-value for a variable 'x' was recorded before a fork in the CFG,
and then we reach the JOIN without 'x' having been modified,
so that the Sym-value assigned to 'x' in all branches is exactly
(physically) the same, then we allow that Sym-value to survive.

Helps PA-2739

test plan:
make test # new tests