refactor: Fully migrate onto new type inference
This absorbs all of the functionality of the old type inference,
operating over `AST_generic.type_`, into the new type inference,
operating over `Type.t`.

The next step will be to do matching directly on `Type.t` and avoid the
conversion back to `AST_generic.type_` for matching. Then, I'd like to
change `id_type` to a `Type.t` so that we represent types with `Type.t`
as much as possible.

This is probably not a pure refactor, but all the tests pass and I did
my best to keep behavior the same. Most likely, some things work better
now, and it's possible that some things work worse. But any regressions
should be an easy fix.

Test plan: Automated tests, including those in pro