feat(interactive): turbo mode (#7896)
## What:
This PR adds in turbo mode for Interactive Mode.

## Why:
It's really friggin cool and nice to have it be instantly responsive

## How:
We use threads to compute the matches concurrently, while taking input
from the user. This allows (relatively)
smooth display of the query bar changing, while not requiring us to
block while computing the matches from the query, which can be done
relatively fast, but not as fast as people can type.

## Problems:
There are a few problems which currently exist
1) The first query is always the longest, due to AST caching. This only
takes about a second or two, but `Lazy.force` is not thread-safe. This
means that if you instantly start typing more than 1 character, threads
will start complaining about being killed because they are trying to
concurrently force the ASTs which are currently being parsed.
2) If you type too fast, Interactive Mode will stall out and stop
updating the display, until your next keypress. This is probably due to
the way that we enter the `Term.event` call, I think that call is
blocking when we should have waited a bit longer for more input, or
something.
3) Originally, I wanted to use the `Term.pending` function, which is
supposed to tell you when `Term.event` is blocking or not. However, I
did a lot of experimentation, and found that it was incredibly
unreliable, and very often would just always tell you that the call
would not block, even when it would. So that's why I resorted to
manually polling `stdin` using `select`.
4) The query bar can lag sometimes

Otherwise, works pretty well.

## Test plan:
https://asciinema.org/a/uyIpst16czMDgVdQT1F2KQNwm

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [ ] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)

---------

Co-authored-by: Yoann Padioleau <pad@r2c.dev>