feat(interactive): ascii art resizing (#7923)
## What:
This PR makes it so the Interactive Mode UI resizes based on the size of
the terminal.

## Why:
Previously, the ASCII art always displayed, and was at a fixed width and
height, meaning it could stretch off the edges if your terminal was too
small. Now, we try to resize to either a smaller ASCII image, or if the
terminal is really really small, we just don't display the ASCII images.

## How:
Added functions which get the ASCII art, which depend on the width and
height. Since the Semgrep logo is wider than it is tall, width is used
as the primary metric for selection. The ghost is taller than it is
wide, so height is used as the primary metric.

## Test plan:
This is what I got by making the terminal as small as it could go
without losing any of the ASCII images:

![image](https://github.com/returntocorp/semgrep/assets/49291449/754a379b-0e52-47ac-ab0e-ece96bc5e73b)


![image](https://github.com/returntocorp/semgrep/assets/49291449/ba0a5a13-0745-4b14-af38-915321ea5330)


![image](https://github.com/returntocorp/semgrep/assets/49291449/c9a144df-72b1-4085-80ad-77948235bbf0)


![image](https://github.com/returntocorp/semgrep/assets/49291449/d0e17ff6-d1da-4cce-9367-94f6d62616f8)

As you can see, there's generous enough whitespace that this should be
good.

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)