js: fix libyaml bindings (#7918)
This PR fixes memory alignment issues that was preventing YAML parsing
from working properly. TL;DR: `ocaml-yaml` uses `ocaml-ctypes` to
generate bindings for the `libyaml` library. Since this is done at
compile time, the memory offsets are calculated based off of the build
machine (64-bit address space --> 8 byte pointer size). WebAssembly,
however, is 32-bit (4 bytes), and this mismatch prevented us from
interfacing with the library properly.

To fix, we override the memory offsets in JavaScript (exploiting its
inherent mutability) at startup. This is a little weird, is much faster
than figuring out how to cross-compile `ocaml-yaml` against a 32-bit
architecture.

There is a weird off-by-one issue with the index of the end token in
Yaml_to_generic, but I'll fix in a separate PR if it proves to be an
issue.

test plan:
- make test
- confirm `js/examples/yaml.html` works properly