fix: Remove special-case for typed-metavar matching
Previously, if a typed metavariable was matched against an identifier,
it would go through the old `m_type` matching codepath rather than the
new type inference codepath. For consistency and simplicity, I have
removed that special case and consolidated on the new type inference and
matching.

This required me to make some improvements to the new type inference and
new matching in order to pass tests. Notably, metavariables inside of
the types in typed metavariables are now bound. Since the AST to which
they are bound is the result of type inference, it is synthetic. I
adopted the pattern pioneered in #6900 of using a local token from the
AST to associate a synthetic AST with a real program location.

The change in test behavior is because we can now match on unresolved
types. The type we infer for `MessageDigest` is just `UnresolvedName
"MessageDigest"` (in a simplified representation). As noted in
`Typing.type_of_name`, this should most likely be `Typeof
"MessageDigest"`, but we conflate class instances with the type of the
class itself throughout our type inference infrastructure, with this
result. I also reformatted that test since I found the mix of tabs and
spaces confusing.

Test plan: Automated tests