feat(Java): Guess types when we can't resolve them definitively
We currently use type inference to handle typed metavariables. For
example, a Java pattern might include `(boolean $X)` which will match
any expression that is inferred to have the type `boolean`.

This PR extends type inference to include heuristics which we use to
make a best guess for types of expressions whose types we cannot infer
using traditional means. This typically happens when there is an
unresolved name involved in the type.

While the heuristics are based on naming patterns in the Java standard
library, they are not limited to the standard library. Nor do we intend
to replicate the entire Java standard library as heuristics. Instead, we
should include heuristics when we believe that they correctly describe
the type of an expression with high confidence, whether they are based
on a convention in the standard library or other common conventions.

Test plan: Automated tests