Fix gomod comment parsing (#7969)
When we parse gomod lockfiles, we look for directives like
`require(...)` and `replace(...)`. We recently added support for
comments in those directives. However, there was a bug where if the
comment ended in a `)`, then the parser interpreted it as the end of the
directive. This fix makes sure that we fully parse through the end of
the comment so we don't misinterpret the `)`, since we know that the
meaningful `)` should not be on a line that starts with `//`.

PR checklist:

- [x] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [x] Tests included or PR comment includes a reproducible test plan
- [x] Documentation is up-to-date
- [x] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [x] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)

---------

Co-authored-by: Matthew McQuaid <matthew@returntocorp.com>