feat(taint): Relate Java getters and setters to properties (#7977)
Previously in Java, Semgrep would consider `o.getX()` tainted if a
tainted argument was passed to `o.setX(...)`. However, it would not
relate that taint back to the underlying property itself. This means
that taint would only be tracked if getters and setters were used in
concert. If a regular assignment to the property was made, followed by a
getter, or if a setter was used, followed by a direct access to the
property, Semgrep would fail to track taint.

Here, I've changed the handling of `o.setX(...)` to make Semgrep
consider the property `o.x` tainted, and added functionality to consider
`o.getX()` tainted if `o.x` is tainted. The modified tests illustrate
the change in behavior.

Test plan: Automated tests

---------

Co-authored-by: Iago Abal <mail@iagoabal.eu>