Revert "feat(taint): Relate Java getters and setters to properties (#7977)" (#7980)
This reverts commit 22b421dc59763f1210077315c8c251168e8aa7a5 (PR #7977).

As pointed out in the review of #7977 this doesn't really help Pro
because the attributes of a class will be resolved and the `sid`s won't
mach. But also we can't deal with getters that way because it doesn't
handle polymorphic taint, e.g. Pro
tests/tainting_interproc/java/get_set_sensitive/ fails.