tainting: Filter parameters taint by type (#7967)
If a parameter has Boolean or number type, and the correspoding option
is enabled, we don't assign any taint to that parameter (polymorphic or
due to a default value).

Also, tweaked `fix_poly_taint_with_field` to check the type of fields,
and except for unresolved Java getter methods, we attach an offset to
polymorphic taint if the offset as a function type (i.e. it is a
method).

Note that I added a typing traversal to infer whether a field is a
method and thus assign it a function type. This conflicted with the
typing heuristics which only triggered when a method had no type
assigned. Decided to move the heuristics to be part of this typing
traversal.

Follows: 64cb387f7a0 ("tainting: Add options to ignore taint based on
types (#7936)")

test plan:
make test # new test