refactor(interactive): kill `Unix.select` (#8007)
## What:
This PR kills our previous workflow of polling `stdin` every `0.0005`
seconds with `Unix.select`, and instead opts for using threads to
determine when events should happen.

## Why:
Ideally, looping but taking actions on events should be able to be done
with the `Term.pending` call, which supposedly tells us when there is a
pending event to the terminal, to be acted upon. Past experimentation
shows that this does not work, so we must find another way.

This way is cleaner, because we don't need to rely on `Unix.select`.

## How:
Instead of `Unix.select`, we spin up another thread (call it the _event
thread_) which simply tightly loops and blocks via `Term.event`. This is
fine, because this just means that the event thread blocks -- the main
thread is free to take input and refresh on an asynchronous flag (such
as the turbo thread completing matches).

## Test plan:
https://asciinema.org/a/VMyjaMUZsTCehnV9783qQHqPS

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)