feat(interactive): buffer matches (#8012)
## What:
This PR makes it so that, instead of only showing matches only once they
are computed for every single file in the workspace, the matches are
buffered in per file as they are computed.

## Why:
This results in an even faster and more tactile experience for the
Interactive Mode user.

## How:
Now, the file zipper for a given state instance is updated in-place
asynchronously by its associated turbo thread. This is precarious,
however, because this means we need to make sure that we don't race when
we write to the zipper for other reasons (for instance, when moving the
file pointer or match pointer around). As such, we have
`atomic_map_file_zipper`, and the file zipper is now protected by a
lock, which is regenerated for each state instance.

I observed that naively making these changes makes the REPL noticeably
slower, such as when you type really fast or backspace a lot. This
seemed to be because we were spawning so many turbo threads even when
there were events left to do, so I made it so we only spawn turbo
threads if there are no pending events that would spawn turbo threads
(which would mean that the turbo thread we spawned was redundant
anyways).

## Test plan:
It's mega fast!!!
https://asciinema.org/a/qyLUJksrqrcDgeGLw8w2NPa6r

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)