taint analysis: Improve handling of dataflow for tainted value propagation in class field definitions (#7975)
This commit resolves an issue where dataflow was not correctly accounted
for when tainted values flowed through field definitions in class
definitions. In JVM-based programming languages such as Java, Kotlin,
and Scala, the order of field definitions is significant, and they
function like statements during object initialization. To address this,
during the transformation from AST to IL, we now treat a list of field
definitions as a sequence of statements, allowing for proper tracking of
tainted values during object initialization.

Closes #7742

test plan:
make test # new test