feat(matching): match record field of ellipsis to any definition (#8112)
## What:
This PR establishes an equivalence between a `FieldDefColon` of an
ellipsis and any other definition kind.

## Why:
`FieldDefColon` of an ellipsis arises from a record field, something
like `{x : ...}` in Javascript.

However, a record might contain a definition which is not a
`FieldDefColon`! A good example of this is
```
const x = foo({
    func: () => {
        return 1;
    },
})
```
which is actually a record of a `FuncDef`. However, based on the
pattern's translation, we should want to match the lambda, even if it's
in a `FuncDef`. So we establish this equivalence.

## How:
Just wrote the case in `Generic_vs_generic`.

## Alternatives considered:
We could alternatively consider establishing a `definition_kind` variant
which itself represents an ellipsis, and translate into that instead of
the original `FieldDefColon`.

I took a poll in Slack, and this seemed to be the better option
(https://semgrepinc.slack.com/archives/C01NXGX2EHZ/p1687384401974879). I
also just like it better. We can reconsider if this comes up more often.

## Test plan:
`make test`

Closes PA-2878

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [X] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)