ci: add polling support for /complete
To ensure that we `semgrep ci` performance continues to scale, we will
be moving some of the processing from /upload_findings_and_ignores into
a background task.

Since the CLI is dependent on some of those processed results, we need
to ensure that the CLI will retry periodically until the scan has been
processed.