chore(osemgrep): send all rules and targets together to Run_semgrep
The run was previously broken up by language, but this is ugly and
unnecessary. I'm refactoring this ahead of changes to implement
osemgrep-pro, which will be easier if this is one command.

(I could also have wrapped it in a function but it wasn't too hard
to fix.)

Test plan: make test