fix(js): Include location of `async` keyword in arrow func range
We don't include the `async` attribute (or any others, if there are
others possible) as part of the generic AST. We might want to change
that, but we might not want to complicate the generic AST for this one
case.

However, since that token does not get included, we get the range wrong.
This is especially a problem for metavariable bindings when they are
used in autofix. See the test case.

We now set `e_range` when constructing the AST. I've used this technique
in a few places in the past to address location issues. To simplify the
logic here, I've also refactored a bit to add a utility which will set
the `e_range` of an expression based on a list of `AST_generic.any`
nodes. Note that this makes it so that calls to `extract_ranges` now
entail one additional allocation for the singleton list. While this is a
hot code path I very much doubt that one more allocation will make a
difference considering all the other work that goes into computing the
ranges.

Fixes #7353

Test plan: Automated tests