Build multi-arch docker images (#8205)
This PR introduces support for multi-arch semgrep docker builds (amd64 +
arm64). A few things to note:
- GHA only provides amd64 runners, so ARM builds are abhorrently slow
([45
mins](https://github.com/returntocorp/semgrep/actions/runs/5456229166/jobs/9928763358)).
I cut us over to depot.dev, which provides dedicated ARM hardware and a
faster docker cache. `depot/build-push-action` will fall back to the
standard docker if depot.dev is offline.
- We run some pre-flight checks prior to pushing the image, which
warranted some tricky maneuvers because neither docker nor depot can
export (and docker can't import) a multi-arch image archive, so instead:
- We build each arch individually, run the checks, and then upload the
archive (name suffixed with the arch)
  - In `push-docker-multi-arch`, we:
    1. Extract the contents of the two images
2. Generate a manifest list so that the docker image will show up as
multi-arch in Docker Hub
    3. Push both images' blobs and manifests to Docker Hub
    4. Push the manifest list for the desired tags to Docker Hub

(this is a bit of a brain-melter so I'm happy to explain further. I can
also add comments to the workflow)

All that being said, we should re-assess if `validate-docker-build.sh`
is still useful (or if we can push to a staging repository), so that we
can rely on docker's/depot's baked-in multi-arch builds

test plan:
- [x] manually triggered `build-test-docker` to push this branch as
`returntocorp/test`. Confirmed that the
[tom-arm64](https://hub.docker.com/r/returntocorp/test/tags) image works
on both amd64 and arm64.

cc: https://github.com/returntocorp/semgrep/issues/2252