tainting: Fix Taint.pick_taint
Given two taints with the same source, Semgrep is supposed to pick the
taint with the shortest call trace, but due to a typo this was not
guaranteed.

Found this by chance while working on a new version of `pick_taint`, I
don't have a test case for it.

Also fixed non-tail-rec use of `List.filter_map` (not sure why we had
not found this before).

test plan:
make test