fix(supply-chain): baseline scans reporting on existing findings
lockfile parsing was caching based only on file name which was resulting
in the lockfile being cached even between baseline and head scans. When
changes were added above an existing vulnerability this would cause
match_based_id to access the wrong lines (everything will be off by new
additions or deletions) which caused deduplication logic to be incorrect

This PR adds a test to catch this case and also caches the parsing on
both filename and last update time of the file.