build aarch64 python wheels (#8243)
This PR adds support for building `aarch64` python wheels. This isn't
simply a copypasta of the `x86_64` workflow because the cost of ARM
emulation in GHA would make this job take forever. Instead, we:
- Add a new intermediate stage to the Dockerfile (`semgrep-wheel`)
which:
1. installs some prereqs (`build-base` because `ruamel.yaml` has native
code that needs to be compiled, `zip` because we gotta zip)
2. copies the arch-specific `semgrep-core` binary from the
`semgrep-core-container` stage
   3. builds the python wheel
4. runs a script to validate that the wheel installs correctly and
semgrep works
(We don't have to worry about folks needlessly building this image
because `docker build` only targets the last image in the Dockerfile by
default)
- Add a new workflow that:
- Kicks off a docker build targeting the `semgrep-wheel` image. This
will be very fast because the docker layer cache will be populated
already
- If the image builds successfully, we extract `dist.zip` from the image
and upload it as an artifact
- Remove the ARM64 check in the CLI (woo!)

If this ends up working smoothly, I'd like to consolidate on using this
workflow for the `x86_64` wheel as well. (future PR)

NOTE: the CLI depends on `ruamel.yaml` which doesn't currently have an
aarch64 python wheel published to PyPI, which means that gcc is required
in order to install this dependency. Not ideal, but better than what we
have currently (which is nothing). Medium-term we can look into
switching from ruamel.yaml to pyyaml, and long-term osemgrep will make
this go away.

test plan:
- checks pass
- grabbed the `manylinux-wheel-aarch64` artifact and confirmed it
installed correctly and semgrep scan worked in an arm64 docker container

related: https://github.com/returntocorp/semgrep/issues/2252