delaying super/self assignment in jsonnet (#8319)
To have self/super be correctly assigned during the plus operator, it is
necessary to delay this assignment until the point where the fields are
accessed (as per the spec). This PR does that, as well as moves the
tests that now pass due to this change into the pass folder.