set proper platform tag for python wheels (#8336)
Backstory: Prior to supporting linux/arm64, we published a semgrep
Python wheel with the [platform compatibility
tag](https://packaging.python.org/en/latest/specifications/platform-compatibility-tags/)
set to `any`, which tells pip to fall back to this if there are no other
platform-specific wheels available.

When we started publishing wheels for both architectures, I naively
assumed that things would Just Work. Unfortunately, the platform
compatibility tags that setup.py was setting by default (`linux_x86_64`
and `linux_aarch64`) are not accepted by PyPI (the reason why is kind of
a long story and has to do with incompatibilities between glibc and
musl).

It turns out the platform compatibility tag is literally just the suffix
on the wheel filename (no changes to the actual content of the wheel),
so this PR basically just updates the generated wheel to have the proper
filenames.

The key word here is filename**s**. In order to support both glibc (e.g.
Debian) and musl (e.g. Alpine) platforms, we need to publish wheels
targeting `manylinux2014` and `musllinux_1_0`, which is why we have the
two-copies-and-a-delete in build-wheels.sh


NOTE: After typing all this out I realized that we should really have
_something_ as the fallback `any` wheel, so I just updated the PR have
`linux_x86_64` wheels remapped to `any`. Doing this ensures that this
change only affects ARM users -- Intel users will keep using the `any`
package as usual

test plan:
- [x] checks pass
- [x] grabbed artifacts from GHA and published to test.pypi.org