fix: move Semgrep LS to OSemgrep (#8321)
Warning: this looks like a big PR but git is lame and did not detect
that a bunch of files were moved and slightly tweaked.

This PR moves the language server from semgrep-core to osemgrep, netting
it a bunch of speed improvements, and overall cleaner code.

## What's actually changed

### Language Server

Primarily, on the python side, we've removed all code relating to
pulling/parsing rules, finding targets, and scan configuration (i.e.
timeout, max target size etc.) this has all been moved to either
Session.ml, or UserSettings.ml.

Because osemgrep provides functionality for getting rules from remote
sources (CI + registry), and targeting, we no longer have to compute
this on the python side -> write a file -> parse it on the core side.
This improves error messaging and handling, speed when scanning new
files/folders, and general stability. Obviously, it decreases complexity
by a ton.

OSemgrep also provides a much better interface for running scans,
meaning we could simplify a lot of code in Diagnostics, post processing,
and in the actual semgrep runner, as we are now using `cli_match`, which
has the rule message w/ metavars rendered, the fix rendered, and some
other things that before we had to compute ourselves. Also, it
handles nosem/nosemgrep tags, which we had to do manually in the LS
before.

Finally, the language server runs inside a LWT runtime (process?), i.e.
inside `Lwt_main.run`, but so do the http helpers for CI/registry
related things. Sadly, sanely, LWT does not allow nested calls to their
runtime, which means that for some config/scan helpers, I had to create
an async version, which is essentially the same as the original version,
but without a final call to the LWT runtime. These methods have a suffix
of _async, and the original methods just call these and pass them to the
runtime to make them synchronous.

### Cohttp

Outside of the Language Server changes listed above, we also switched
our http client to cohttp. This is because the previous http client
would randomly fail on DNS request timeouts. Also, cohttp provides jsoo
bindings which will be useful in the future.

## Reorganization

For the most part, the organization is the same, except that we finally
split out the JSONRPC logic of the language server from the
semgrep/message handling. The RPCServer.ml file is existing code, just
split out. This did not have to be a part of this PR, but I figured I'd
do it while I'm already moving things around. The goal was to separate
any generic JSONRPC/LSP logic, from Semgrep LS logic, resulting in
cleaner code, that should be less confusing for future contributors.

## Testing

Most of the testing remains the same, and all passes, except testing
related to CI/logging in. These tests previously relied on the fact that
all CI rules were pulled on the python side, so we could easily mock
them out. Now the CI logic is in the language server, and there's not
much testing infra there, especially to mock this out. I decided to just
remove these tests and test them manually. I'm also ok with this as we
are now using logic from osemgrep to get rules, and those (should) be
already tested, so I'm ok leaning on that too.

PR checklist:

- [X] Purpose of the code is [evident to future
readers](https://semgrep.dev/docs/contributing/contributing-code/#explaining-code)
- [X] Tests included or PR comment includes a reproducible test plan
- [X] Documentation is up-to-date
- [ ] A changelog entry was [added to
changelog.d](https://semgrep.dev/docs/contributing/contributing-code/#adding-a-changelog-entry)
for any user-facing change
- [X] Change has no security implications (otherwise, ping security
team)

If you're unsure about any of this, please see:

- [Contribution
guidelines](https://semgrep.dev/docs/contributing/contributing-code)!
- [One of the more specific guides located
here](https://semgrep.dev/docs/contributing/contributing/)

---------

Co-authored-by: semgrep-app[bot] <63493438+semgrep-app[bot]@users.noreply.github.com>