refactor: Generalize id_hidden to id_flags
We can store more flags in the same space by using an `int` as a
bit-vector instead of individual `bool` fields. (I am planning to
add an extra flag to record whether a `VarDef' is constant, which
is useful for taint analysis.)

Also, made `id_flags' a ref so it can be updated.

test plan:
make test