Use env model when handling std library functions to speed up evaluation in jsonnet (#8423)
In the substitution model of eval, we just always pass an empty
environment now, so that we can use the same value file for both. With
that, we can then call the environment model evaluation for std library
functions so that they don't take as long. We want to use substitution
model for other things since it implements more features correctly, but
since std library does not use feature that aren't implemented correctly
in the environment model, it is safe to revert to it just for std
library.