feat: implicit return support (for ruby and julia)
The implementation revolves around 3 new fields added to `tin`
* inside_function
* parent_is_last_stmt_trans
* self_is_last_stmt

`inside_function` is initialized to false and becomes true whenever a
function is visited.

`parent_is_last_stmt_trans` is initialized to true and when visiting
each substmt, it tells the substmt whether itself upto root is true.

`self_is_last_stmt` is initialized to true except for blocks. When we
enter a block, it is set to false, except for only the last statement
in that block which is set to true.

This is one of the tasks needed to fix #8408.

Added new tests and tested for regressions with `make core-test`.