matching: ellipsis: Always try going deep (#8440)
Pattern `... bar()` would behave strangely because it would only try to
match deeply if it did not match anything non-deeply. So depending on
how the target code looked, `... bar()` would match or not match inside
e.g. an `if` statement. This was confusing and not a great property for
Semgrep to have.

The regression was introduced in PR #852 as part of a set of
optimizations done back in 0.9.0 (!), but at present reverting this one
thing does not seem to have any negative perf impact.

Follows: c1ca429c183 ("Optimize deep statement matching (#852)")

Closes returntocorp/semgrep-rules#660
Closes PA-2992

test plan:
make test # new tests
Also, compared against develop running p/default on 32 repos from
stress-test-monorepo, and no meaningful slowdown or increase in memory
usage was observed.