osemgrep: switching to real e2e testing, part 1
our end-to-end (e2e) tests ran by 'make e2e' are not
fully end-to-end because they often don't call semgrep
but instead use an optimization hack called the ClickRunner
to avoid a fork. To be real e2e, we need the fork instead,
and we need to call cli/bin/semgrep, especially if we want
to port more code to osemgrep; otherwise the test will just
exercise the pysemgrep code

This is just the first part of the process of moving away
from the ClickRunner to a real fork.

test plan:
make e2e