remove semgrep shouldafound
The command was not really used (1 issue reported per month
in https://github.com/returntocorp/issue-test/issues and
in Slack under #shouldafound-alerts).
It's just simpler to remove rather than spending the time to port
it to osemgrep.

This mostly revert #5291

https://linear.app/semgrep/issue/PA-3003/create-a-shouldafound-alternative-in-the-vscode-extension
is tracking moving such a feature to our vscode extension instead.

BTW the backend code is in
https://github.com/returntocorp/shouldafound
and the only occurence in semgrep-app is in an error message (that
I'll update in a separate PR).

test plan:
make test