osemgrep: get rid of command.py
This class seems an overcomplicated solution for just checking whether
we ran from semgrep ci. Maybe it was not mean to be extended
and further used, but it's just used for one thing right now,
so let's just simplify the code and get rid of it.

test plan:
make e2e