osemgrep: switch to real e2e testing, part 2
This is a followup to #8492.
Our end-to-end (e2e) tests ran by 'make e2e' were not
fully end-to-end because they often didn't call semgrep
but instead use an optimization hack called the ClickRunner
to avoid a fork.
With this PR, we switch back to a real fork, which is slower,
but correct.
There's only a few remaining e2e tests that still use the
Click_runner (mostly test_ci.py because of heavy use of Python
mocking).

test plan:
make e2e!