refactor: Separate partitions for Gensym
As noted in the comments, this allows us to generate a data structure
that uses IDs, serialize  it, and then load it in a future run without
fear of collisions.

Specifically, I plan to use this for library definitions in the Pro
Engine, where I will create a SIG using partition B for sids, serialize
it, and then load it. Because normal runs use partition A, the sids
generated for the libdef SIG will not collide.

Test plan: Existing automated tests ensure that partition A continues to
work as expected. I have also tested with my work in progress on
libdefs, which are generated with partition B, and all tests pass.